soma2

+++ soma.c
@@ -1,10 +1,10 @@
 
 int main (){
-    float valor1, valor2, soma;
+    int valor1, valor2, soma;
     printf("Digite o primeiro valor:\n");
-    scanf("%f", &valor1);
+    scanf("%d", &valor1);
     printf("Digite o segundo valor:\n");
-    scanf("%f", &valor2);
+    scanf("%d", &valor2);
     soma=valor1+valor2;
-    printf("A soma dos dois termos é: %f", soma);
+    printf("A soma dos dois termos é: %d", soma);
 }
